refactor(sections): move shared section constants and validation out of programs
