+++ src/components/common/section-templates/dual-images-bottom/DualImagesBottom.tsx
@@ -1,7 +1,7 @@
 import cn from 'classnames';
 import { ImagesBottomSection } from '../shared/images-bottom-section/ImagesBottomSection';
 import { ImageValues, Image } from '@/types/common/image';
-import { DUAL_IMAGES_CONFIG } from '@/const/admin/programs';
+import { DUAL_IMAGES_CONFIG } from '@/const/admin/sections';
 import { SectionMode, SectionTemplate } from '@/types/common/sections';
 import styles from './DualImagesBottom.module.scss';
 import viewStyles from './ViewDualImagesBottom.module.scss';
