+++ src/components/common/section-templates/triple-images-bottom/TripleImagesBottom.tsx
@@ -1,7 +1,7 @@
 import cn from 'classnames';
 import { ImagesBottomSection } from '../shared/images-bottom-section/ImagesBottomSection';
 import { ImageValues, Image } from '@/types/common/image';
-import { TRIPLE_IMAGES_CONFIG } from '@/const/admin/programs';
+import { TRIPLE_IMAGES_CONFIG } from '@/const/admin/sections';
 import { SectionMode, SectionTemplate } from '@/types/common/sections';
 import styles from './TripleImagesBottom.module.scss';
 import viewStyles from './ViewTripleImagesBottom.module.scss';
