+++ src/pages/admin/programs/components/programs-page-modals/add-section-modal/test-utils/addSectionModalTestUtils.ts
@@ -3,6 +3,7 @@ import { render, screen, fireEvent } from '@testing-library/react';
 import { AddSectionModal } from '../AddSectionModal';
 import type { AddSectionModalProps } from '../AddSectionModal';
 import { PROGRAMS_TEXT } from '@/const/admin/programs';
+import { SECTIONS_TEXT } from '@/const/admin/sections';
 
 export const renderAddSectionModal = (props: AddSectionModalProps) => {
     render(React.createElement(AddSectionModal, props));
@@ -31,7 +32,7 @@ export const clickNextButton = () => fireEvent.click(getNextButton());
 export const clickCloseButton = () => fireEvent.click(getCloseButton());
 
 export const buildFiveShortDescriptions = () =>
-    Array.from({ length: 5 }, () => PROGRAMS_TEXT.SECTION.DESCRIPTION_SAMPLE_TEXT_SHORT);
+    Array.from({ length: 5 }, () => SECTIONS_TEXT.SECTION.DESCRIPTION_SAMPLE_TEXT_SHORT);
 
 export const findFirstCallByTemplateId = (mockFn: jest.Mock, templateId: any) => {
     const calls = mockFn.mock.calls.map((call) => call[0]);
